trails-viz-api/requirements.txt
Flask==1.1.1
numpy==1.16.4
pandas==0.25.1
geopandas==0.4.1
flask-cors==3.0.8
Flask-Bcrypt==0.7.1
google-cloud-storage==1.20.0
